update: refactor expense service to use the new shared common library
