+++ expense-service/src/main/java/com/huiun/fizzybudget/expenseservice/config/AppConfig.java
@@ -1,4 +1,4 @@
-package com.huiun.fizzybudget.userservice.config;
+package com.huiun.fizzybudget.expenseservice.config;
 
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
