+++ expense-service/src/main/java/com/huiun/fizzybudget/expenseservice/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.huiun.fizzybudget.expenseservice.repository;
 
-import com.huiun.fizzybudget.sharedentities.Category;
+import com.huiun.fizzybudget.common.entities.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
