+++ expense-service/src/main/java/com/huiun/fizzybudget/expenseservice/repository/CurrencyRepository.java
@@ -1,6 +1,6 @@
 package com.huiun.fizzybudget.expenseservice.repository;
 
-import com.huiun.fizzybudget.sharedentities.Currency;
+import com.huiun.fizzybudget.common.entities.Currency;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CurrencyRepository extends JpaRepository<Currency, Long> {
