+++ expense-service/src/main/java/com/huiun/fizzybudget/expenseservice/repository/ExpenseRepository.java
@@ -1,6 +1,6 @@
 package com.huiun.fizzybudget.expenseservice.repository;
 
-import com.huiun.fizzybudget.sharedentities.Expense;
+import com.huiun.fizzybudget.common.entities.Expense;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
