+++ expense-service/src/main/java/com/huiun/fizzybudget/expenseservice/repository/RecurrentExpenseRepository.java
@@ -1,6 +1,6 @@
 package com.huiun.fizzybudget.expenseservice.repository;
 
-import com.huiun.fizzybudget.sharedentities.RecurrentExpense;
+import com.huiun.fizzybudget.common.entities.RecurrentExpense;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RecurrentExpenseRepository extends JpaRepository<RecurrentExpense, Long> {
